refactor: moved login and mainscene to containers
